This is after what we done in 8.12

+++ src/main/java/bgu/spl/mics/CallbackDeactivation.java
@@ -10,7 +10,7 @@ public class CallbackDeactivation  implements Callback {
     @Override
     public void call(Object c) {
         R2D2Microservice tempobject=(R2D2Microservice) c ;
-        int n= tempobject.getDuration();
+        long n= tempobject.getDuration();
         try {
             wait(n);
         } catch (InterruptedException e) {
